+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Shows aggregated and self time of projects in report (#33)
 * Projects have a color (#35)
 * Week report with vertical day columns and projects denoted by color and initial letter: `track report week [DATE]` (#35)
+* Auto-scale day and week report width based on terminal width (#35)
 
 ### Bug fixes